+++ vcd/config_test.go
@@ -813,3 +813,21 @@ func importStateIdEdgeGatewayObject(vcd TestConfig, edgeGatewayName, objectName 
 			objectName, nil
 	}
 }
+
+// Used by all entities that depend on Org + VDC + vApp VM (such as VM internal disks)
+func importStateIdVmObject(orgName, vdcName, vappName, vmName, objectIdentifier string) resource.ImportStateIdFunc {
+	return func(*terraform.State) (string, error) {
+		if orgName == "" || vdcName == "" || vappName == "" || vmName == "" || objectIdentifier == "" {
+			return "", fmt.Errorf("missing information to generate import path")
+		}
+		return orgName +
+			ImportSeparator +
+			vdcName +
+			ImportSeparator +
+			vappName +
+			ImportSeparator +
+			vmName +
+			ImportSeparator +
+			objectIdentifier, nil
+	}
+}